+++ src/cbadc/digital_estimator/iir_estimator.py
@@ -152,9 +152,7 @@ class IIRFilter(BatchEstimator):
         for k2 in range(self.K2):
             self.h[:, k2, :] = np.dot(self.WT, temp2)
             temp2 = np.dot(self.Ab, temp2)
-        self._control_signal_valued = np.zeros(
-            (self.K2, self.analog_system.M), dtype=np.int8
-        )
+        self._control_signal_valued = np.zeros((self.K2, self.analog_system.M))
         self._mean = np.zeros(self.analog_system.N, dtype=np.double)
 
     def __iter__(self):
@@ -180,7 +178,7 @@ class IIRFilter(BatchEstimator):
             logger.warning("Estimator received Stop Iteration")
             raise StopIteration
 
-        self._control_signal_valued[-1, :] = np.asarray(2 * temp - 1, dtype=np.int8)
+        self._control_signal_valued[-1, :] = np.asarray(2 * temp - 1)
 
         # self._control_signal_valued.shape -> (K2, M)
         # self.h.shape -> (L, K2, M)
