+++ src/cbadc/digital_estimator/parallel_digital_estimator.py
@@ -195,7 +195,7 @@ class ParallelEstimator(BatchEstimator):
 
     def _allocate_memory_buffers(self):
         # Allocate memory buffers
-        self._control_signal = np.zeros((self.K3, self.analog_system.M), dtype=np.int8)
+        self._control_signal = np.zeros((self.K3, self.analog_system.M))
         self._estimate = np.zeros((self.K1, self.analog_system.L), dtype=np.double)
         self._control_signal_in_buffer = 0
         self._mean = np.zeros((self.analog_system.N), dtype=np.complex128)
@@ -240,9 +240,7 @@ class ParallelEstimator(BatchEstimator):
             raise Exception(
                 "Input buffer full. You must compute batch before adding more control signals"
             )
-        self._control_signal[self._control_signal_in_buffer, :] = np.asarray(
-            s, dtype=np.int8
-        )
+        self._control_signal[self._control_signal_in_buffer, :] = np.asarray(s)
         self._control_signal_in_buffer += 1
         return self._control_signal_in_buffer > (self.K3 - 1)
 
@@ -280,7 +278,7 @@ class ParallelEstimator(BatchEstimator):
                 control_signal_sample = next(self.control_signal)
             except RuntimeError:
                 self._stop_iteration = True
-                control_signal_sample = np.zeros((self.analog_system.M), dtype=np.int8)
+                control_signal_sample = np.zeros((self.analog_system.M))
             full = self._input(control_signal_sample)
 
         # Compute new batch of K1 estimates
